chore: move most of the implementation to providers module
